~переработанный id +history db

+++ src/ToDoProject/Abstractions/IStorage.java
@@ -6,6 +6,6 @@ import java.io.IOException;
 public interface IStorage {
     Task[] getAll () throws IOException;
     void add (Task data) throws IOException;
-    void delete (int id) throws IOException;
-    void setStatus (int id, String status) throws IOException;
+    void delete (long id) throws IOException;
+    void setStatus (long id, String status) throws IOException;
 }
